Update getting-started.md to install as dependencies
I've updated to install reanimated and gesturehandler as dependencies since they will be used in production code.

## Uncertainties 
Is `wcag-color` even needed to be installed by consumers of this package if it is a peerDependency of the `react-native-ama`. My understanding is that this can be removed but I may be missing something.